[aot] Add task_type for OpenGL
