+++ taichi/backends/opengl/opengl_api.cpp
@@ -1,25 +1,25 @@
-//#define _GLSL_DEBUG 1
 #include "opengl_api.h"
 
+#include <list>
+
 #include "taichi/backends/opengl/opengl_kernel_util.h"
+#include "taichi/backends/opengl/opengl_utils.h"
+#include "taichi/backends/opengl/shaders/runtime.h"
+#include "taichi/ir/transforms.h"
 #include "taichi/program/kernel.h"
 #include "taichi/program/program.h"
 #include "taichi/program/py_print_buffer.h"
 #include "taichi/util/environ_config.h"
-#include "taichi/backends/opengl/shaders/runtime.h"
-#include "taichi/ir/transforms.h"
-#include "taichi/backends/opengl/opengl_utils.h"
 
 #ifdef TI_WITH_OPENGL
-#include "glad/gl.h"
-#include "glad/egl.h"
 #include "GLFW/glfw3.h"
+#include "glad/egl.h"
+#include "glad/gl.h"
 #include "taichi/backends/opengl/opengl_device.h"
-#endif
+#endif  // TI_WITH_OPENGL
 
-#include <list>
-
-TLANG_NAMESPACE_BEGIN
+namespace taichi {
+namespace lang {
 namespace opengl {
 
 #define PER_OPENGL_EXTENSION(x) bool opengl_extension_##x;
@@ -253,6 +253,7 @@ void CompiledTaichiKernel::init_args(Kernel *kernel) {
 void CompiledTaichiKernel::add(
     const std::string &name,
     const std::string &source_code,
+    OffloadedTaskType type,
     int num_workgroups,
     int workgroup_size,
     std::unordered_map<int, irpass::ExternalPtrAccess> *ext_ptr_access) {
@@ -271,7 +272,7 @@ void CompiledTaichiKernel::add(
 
   TI_DEBUG("[glsl]\ncompiling kernel {}<<<{}, {}>>>:\n{}", name, num_workgroups,
            workgroup_size, source);
-  tasks.push_back({name, source, workgroup_size, num_workgroups});
+  tasks.push_back({name, source, type, workgroup_size, num_workgroups});
 
   if (ext_ptr_access) {
     for (auto pair : *ext_ptr_access) {
@@ -609,4 +610,5 @@ bool is_gles() {
 }
 
 }  // namespace opengl
-TLANG_NAMESPACE_END
+}  // namespace lang
+}  // namespace taichi
