+++ taichi/backends/opengl/opengl_api.h
@@ -1,20 +1,21 @@
 #pragma once
 
-#include "taichi/common/core.h"
-#include "taichi/ir/transforms.h"
-
+#include <optional>
 #include <string>
 #include <vector>
-#include <optional>
 
 #include "taichi/backends/device.h"
-#include "taichi/backends/opengl/opengl_kernel_util.h"
 #include "taichi/backends/opengl/opengl_kernel_launcher.h"
+#include "taichi/backends/opengl/opengl_kernel_util.h"
+#include "taichi/common/core.h"
+#include "taichi/ir/offloaded_task_type.h"
+#include "taichi/ir/transforms.h"
 #define TI_RUNTIME_HOST
 #include "taichi/program/context.h"
 #undef TI_RUNTIME_HOST
 
-TLANG_NAMESPACE_BEGIN
+namespace taichi {
+namespace lang {
 
 class Kernel;
 class OffloadedStmt;
@@ -43,6 +44,7 @@ extern int opengl_threads_per_block;
 struct CompiledOffloadedTask {
   std::string name;
   std::string src;
+  OffloadedTaskType type;
   int workgroup_size;
   int num_groups;
 
@@ -74,6 +76,7 @@ struct CompiledTaichiKernel {
   void init_args(Kernel *kernel);
   void add(const std::string &name,
            const std::string &source_code,
+           OffloadedTaskType type,
            int num_workgrous,
            int workgroup_size,
            std::unordered_map<int, irpass::ExternalPtrAccess> *ext_ptr_access =
@@ -130,5 +133,5 @@ class DeviceCompiledTaichiKernel {
 };
 
 }  // namespace opengl
-
-TLANG_NAMESPACE_END
+}  // namespace lang
+}  // namespace taichi
